new approch

+++ src/utils/time.c
@@ -6,13 +6,13 @@
 /*   By: mohammoh <mohammoh@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/20 18:03:47 by mohammoh          #+#    #+#             */
-/*   Updated: 2024/05/22 15:20:36 by mohammoh         ###   ########.fr       */
+/*   Updated: 2024/05/23 19:24:47 by mohammoh         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/philo.h"
 
-long	ft_time(void)
+size_t	ft_time(void)
 {
 	struct timeval	time;
 
@@ -20,12 +20,13 @@ long	ft_time(void)
 	return ((time.tv_sec * 1000) + (time.tv_usec / 1000));
 }
 
-void	ft_usleep(int ms)
+void	ft_usleep(size_t ms)
 {
 
-	int	start;
+	size_t	start;
 
+	//check if his time to die has finished then end the simulation
 	start = ft_time();
-	while ((ft_time() - start) < ms)
+	while ((ft_time() - start) < ms )
 		usleep(500);
 }
